refactor: lintで出ていたproblemsを修正
