+++ lib/app.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:lifecycle/lifecycle.dart';
 import 'package:mybt/flavors.dart';
-import 'package:mybt/res/R.dart';
+import 'package:mybt/res/res.dart';
 import 'package:mybt/ui/start/splash_page.dart';
 
 class MyApp extends StatelessWidget {
