+++ lib/repository/local/local_data_source.dart
@@ -1,9 +1,8 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:mybt/repository/local/entities/history_entity.dart';
 
-final localDataSourceProvider = Provider((ref) => _LocalDataSource());
+final localDataSourceProvider = Provider((ref) => const _LocalDataSource());
 
 class _LocalDataSource {
   const _LocalDataSource();
