+++ lib/repository/local/setting_dao.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-final settingDaoProvider = Provider((ref) => SettingDao());
+final settingDaoProvider = Provider((ref) => const SettingDao());
 
 class SettingDao {
   const SettingDao();
