+++ lib/ui/start/splash_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:mybt/models/app_setting.dart';
-import 'package:mybt/res/R.dart';
+import 'package:mybt/res/res.dart';
 import 'package:mybt/ui/home/home_page.dart';
 import 'package:mybt/ui/start/splash_view_model.dart';
 import 'package:mybt/ui/start/start_page.dart';
